+++ activerecord/test/cases/enum_test.rb
@@ -124,11 +124,16 @@ class EnumTest < ActiveRecord::TestCase
     assert_predicate @book, :written?
   end
 
+  test "method assignment" do
+    @book.set_written
+    assert_predicate @book, :written?
+  end
+
   test "enum changed attributes" do
     old_status = @book.status
     old_language = @book.language
     @book.status = :proposed
-    @book.language = :spanish
+    @book.set_spanish
     assert_equal old_status, @book.changed_attributes[:status]
     assert_equal old_language, @book.changed_attributes[:language]
   end
@@ -147,7 +152,7 @@ class EnumTest < ActiveRecord::TestCase
     old_status = @book.status
     old_language = @book.language
     @book.status = :proposed
-    @book.language = :spanish
+    @book.set_spanish
     assert_equal [old_status, "proposed"], @book.changes[:status]
     assert_equal [old_language, "spanish"], @book.changes[:language]
   end
@@ -156,21 +161,21 @@ class EnumTest < ActiveRecord::TestCase
     old_status = @book.status
     old_language = @book.language
     @book.status = :published
-    @book.language = :spanish
+    @book.set_spanish
     assert_equal old_status, @book.attribute_was(:status)
     assert_equal old_language, @book.attribute_was(:language)
   end
 
   test "enum attribute changed" do
     @book.status = :proposed
-    @book.language = :french
+    @book.set_french
     assert @book.attribute_changed?(:status)
     assert @book.attribute_changed?(:language)
   end
 
   test "enum attribute changed to" do
     @book.status = :proposed
-    @book.language = :french
+    @book.set_french
     assert @book.attribute_changed?(:status, to: "proposed")
     assert @book.attribute_changed?(:language, to: "french")
   end
@@ -179,7 +184,7 @@ class EnumTest < ActiveRecord::TestCase
     old_status = @book.status
     old_language = @book.language
     @book.status = :proposed
-    @book.language = :french
+    @book.set_french
     assert @book.attribute_changed?(:status, from: old_status)
     assert @book.attribute_changed?(:language, from: old_language)
   end
@@ -188,7 +193,7 @@ class EnumTest < ActiveRecord::TestCase
     old_status = @book.status
     old_language = @book.language
     @book.status = :proposed
-    @book.language = :french
+    @book.set_french
     assert @book.attribute_changed?(:status, from: old_status, to: "proposed")
     assert @book.attribute_changed?(:language, from: old_language, to: "french")
   end
@@ -202,7 +207,7 @@ class EnumTest < ActiveRecord::TestCase
   test "persist changes that are dirty" do
     @book.status = :proposed
     assert @book.attribute_changed?(:status)
-    @book.status = :written
+    @book.set_written
     assert @book.attribute_changed?(:status)
   end